Allow logreader to read var_log_t
It's not the default on CentOS, but if persistent storage is enabled for
journald by setting Storage=persistent in /etc/systemd/journald.conf,
the log files will have the same label as the rest of /var/log, which is
var_log_t. This kind of journald configuration is valid and supported by
the logging chart, so add policy rules for both CentOS 7 and CentOS 8 to
allow the rke_logreader_t domain to read this files.